deconstructing table elements into react components; avoids setting an arbitrary variable called 'component' in order to access component functions - using 'this' instead
